py/requirements.txt
yt-dlp
Pillow
mutagen
